refactor(infra): extract bench-suite provenance helpers
Move the git and package provenance helpers out of run.py into
bench_suite/provenance.py so the same code can run inside a remote
benchmark worker (no git binary), and add worker-side identity helpers
(git HEAD without git, content fingerprints, TVM identity, loaded
libraries).
